Create README.md
customHarvester.py is a harvester for the CKAN data management system which was developed for the Cypriot open data portal http://www.data.gov.cy/ in 2016. This custom harvester was implemented as a part of a senior project.

The video below demonstrates the CKAN instance that was developed. Extensions like ckanext-geoview, ckanext-spatial and ckanext-harvest were integrated since the purpose of the system was to "harvest" datasets from the national portal.
https://www.youtube.com/watch?v=EtvVwfUlZPg